requirements.txt
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.1.0
joblib>=1.2.0
flask>=2.3.0
beautifulsoup4>=4.11.0
requests>=2.28.0
Pillow>=9.0.0
pytesseract>=0.3.10
lxml>=4.9.0
reportlab>=4.0.0
python-dateutil>=2.8.0 